+++ docs/src/release-notes.md
@@ -14,11 +14,11 @@ The format of the release notes follows [Keep a Changelog](https://keepachangelo
 - Improved error handling during download
 - Methods `convert`/`convert!` for batch conversions of inventory data files separated from
   download sessions ([#16](https://github.com/LIM-AeroCloud/ICARE.jl/issues/16))
-- Method `clean` to clean the product folder recursively from any files not belonging to the inventory
-  ([#16](https://github.com/LIM-AeroCloud/ICARE.jl/issues/16))
+- Method `clean!` to clean the product folder recursively from any files not belonging to the
+  inventory ([#16](https://github.com/LIM-AeroCloud/ICARE.jl/issues/16))
 - Functions `ignore!` and `unignore!` to exclude database files in the inventory and include
   them again ([#17](https://github.com/LIM-AeroCloud/ICARE.jl/issues/17))
-- Functions `attach!` and `detach!` to mark extra data exempt from `clean` procedures ([#17](https://github.com/LIM-AeroCloud/ICARE.jl/issues/17))
+- Functions `attach!` and `detach!` to mark extra data exempt from `clean!` procedures ([#17](https://github.com/LIM-AeroCloud/ICARE.jl/issues/17))
 - Function `list_inventory` to show a simplified folder and file tree and statistics about
   the inventory and the downloaded portion ([#17](https://github.com/LIM-AeroCloud/ICARE.jl/issues/17))